delet old frontend
